reorganise jba
