+++ .gitignore
@@ -31,3 +31,8 @@ JBA/Nepal/E151185/combined/model_data/footprint.csv
 JBA/Nepal/E151185/combined/keys_data/areaperil_dict.csv
 JBA/Nepal/E151185/combined/keys_data/areaperil_dict.parquet
 JBA/Nepal/E432557/combined/keys_data/areaperil_dict.csv
+JBA/Philippines-Taiwan/combined-undefended/model_data/footprint.csv
+JBA/Philippines-Taiwan/combined-undefended/keys_data/areaperil_dict.csv
+JBA/Philippines-Taiwan/combined-undefended/keys_data/areaperil_dict.parquet
+JBA/Philippines-Taiwan/combined-defended/model_data/footprint.csv
+JBA/Philippines-Taiwan/combined-defended/keys_data/areaperil_dict.csv
