+++ JBA/batch_configs/Ghana/configE635900_FLSW.json
