+++ JBA/batch_configs/Ghana/configE689982_FLSW.json
