--- JBA/s3upload.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-# aws s3 sync ./Ghana s3://oasislmf-scenarios/JBA/Ghana
-aws s3 sync ./Nepal s3://oasislmf-scenarios/JBA/Nepal
-aws s3 sync ./Philippines-Taiwan s3://oasislmf-scenarios/JBA/Philippines-Taiwan
